requirements.txt
pytest~=7.4.2
selene==2.0.0rc4
allure-pytest
pydantic-settings
selenium~=4.14.0
appium-python-client~=3.1.1